Separated out control system tests from load tests.
